PROBLEM with my docker compose file , as far as I was saving my Azure Secret keys there
Chnaged them

+++ Services/Ordering/Ordering.Domain/Models/OrderItem.cs
@@ -1,7 +1,7 @@
 namespace Ordering.Domain.Models;
 public class OrderItem : Entity<OrderItemId>
 {
-    internal OrderItem(OrderId orderId,ProductId productId, decimal price, decimal quantity)
+    internal OrderItem(OrderId orderId,ProductId productId, decimal price, int quantity)
     {
         Id = OrderItemId.Of(Guid.NewGuid());
         OrderId = orderId;
@@ -13,5 +13,5 @@ public class OrderItem : Entity<OrderItemId>
     public OrderId OrderId { get; private set; } = default!;
     public ProductId ProductId { get; private set; } = default!;
     public decimal Price { get; private set; } = default!;
-    public decimal Quantity { get; private set; } = default!;
+    public int Quantity { get; private set; } = default!;
 }
